% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{adjust_path}
\alias{adjust_path}
\title{Adjust image path}
\usage{
adjust_path(path_to_child_folder, rp)
}
\arguments{
\item{path_to_child_folder}{(\code{character(1)}) \cr
Path to child folder.}

\item{rp}{???}
}
\description{
Adjust the img path of a chapter
}
