% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accesors.R
\name{getPackagesFromSlides}
\alias{getPackagesFromSlides}
\title{Get package names from slideset.}
\usage{
getPackagesFromSlides(course.list)
}
\arguments{
\item{course.list}{\code{\link{list}}\cr
The list object containing the course structure.}
}
\description{
This function looks for packages used within the slides. However, improvement is needed for a nice List of the packages.
}
