% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accesors.R
\name{getSlidePaths}
\alias{getSlidePaths}
\title{Get directory of used slidesets.}
\usage{
getSlidePaths(course_list)
}
\arguments{
\item{course_list}{\code{\link{list}}\cr
The list object containing the course structure.}
}
\description{
This function returns a data frame containing the path to the file location of a specific slide chunk.
}
