% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structure.R
\name{getStructure}
\alias{getStructure}
\title{Get the structure of all existing chapters}
\usage{
getStructure(type = "chapters", path = getwd())
}
\arguments{
\item{type}{(\code{character(1)})\cr
The first subdirectory of the files. Have to be \code{chapters} or
\code{exercises}.}

\item{path}{(\code{character(1)})\cr
The working directory of the rcourses repo. Please notice, that every path
containing \code{/rcourses} will work.}
}
\value{
\link[data.table:data.table]{data.table::data.table}\cr
\code{getStructure} returns a \code{data.table} with all the existing
chapter or exercise files. The data table contains the following variables:
- course:   Course which contains the file
- chapters: Chapter which contains the file
- files:    Name of the files
}
\description{
The function \code{getStructure} reads in the structure of all existing
Rmd chapters/exercises in the chapters or exercises directory. This file
is also needed in the \code{setup.Rmd} file. In the most cases this
function is used to get the specific directory a file lies in. Therefore
the chapter names have to be unique!

The generation of the structure is done automatically. This is very handy
if someone copies Rmd files from one chapter or course to another.

The \code{files} column in the resulting data.table is also the name one
can use to create a course list and create a course Rmd file with
\code{createCourse}.
}
