% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nExercises}
\alias{nExercises}
\title{Create an exercise slide.}
\usage{
nExercises(n)
}
\arguments{
\item{n}{\code{\link{integer(1)}}\cr
The number of exercises shown on the slide.}
}
\value{
\code{\link{character(2)}}\cr
A character vector containing the lines of the exercise slide. This vector
has the class \code{exercise_slide} which is needed for the function
\code{createCourse}, that it recognize how to deal with the character
vector.
}
\description{
While creating the list with the course structure, with the \code{nExercise}
function it is very easy to include slides for the exercises. The only
parameter is the number of exercise which should be shown on the slide. If
you want a slide with three exercises on it, than just use \code{nExercise(3)}.
This function is only implemented for 0 < n < 4, and creates for \code{n = 3}
the following slide:\preformatted{#
\threeExercises
}

The main advantage of using this function is the counter used in the
latex command \code{threeExercises}. Notice, that the counter will be
resetted after every subsection!
}
