% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_rmd.R
\name{openRmd}
\alias{openRmd}
\title{Short function to open a Rmd file}
\usage{
openRmd(rmd.file)
}
\arguments{
\item{rmd.file}{(\code{character(1)})\cr
Slide name.}
}
\description{
This function opens a Rmd file. Please use just file names like they are
named in the getStructure function
}
