% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_file.R
\name{renderFile}
\alias{renderFile}
\title{Render just one slide set.}
\usage{
renderFile(slide.name, open.pdf = FALSE, ...)
}
\arguments{
\item{slide.name}{(\code{character(1)})\cr
Name of the slide as given from the getStructure function.}

\item{open.pdf}{(\code{character(1)})\cr
A character vector telling whether the resulting PDF file should be opened. Default = FALSE.}

\item{...}{Passed to \code{createCourse()}}
}
\value{
\code{\link{character}}\cr
A character vector containing the lines of the final Rmd file for the
specific file named in 'slide.name'.
}
\description{
This function creates and render just one slide set.
}
