% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{slide}
\alias{slide}
\title{Create an customized slide.}
\usage{
slide(...)
}
\arguments{
\item{...}{\code{\link{character}}\cr
Arguments passed to \code{list}. This list will then converted to a vector.
Therefore be careful with type conversions. To be on the save side, just
use character vectors.}
}
\value{
\code{\link{character}}\cr
A character vector containing the lines of the slide. This vector
has the class \code{empty_slide} which is needed for the function
\code{createCourse}, that it recognize how to deal with the character
vector.
}
\description{
While creating the list with the course structure, with the \code{slide}
function it is easy to include a slide with customized content. Please
use \code{\\\\} instead of \code{\\} for latex commands (see example
below).
}
\examples{
  test_slide = slide("# Test Slide",
    "The euclidean norm is defined by:",
    "\\\\[ \\\\|x\\\\|_2 = \\\\left( \\\\sum\\\\limits_{i = 1}^n x_i^2 \\\\right)^{1 / 2} \\\\]")

  writeLines(test_slide)
}
